requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
narwhals==1.30.0
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
pymongo==4.11.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
rsa==4.9
six==1.17.0
smmap==5.0.2
streamlit==1.43.2
streamlit-aggrid==1.1.2
streamlit-drawable-canvas==0.9.3
streamlit-option-menu==0.4.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
